Make it visible in all pages by putting it in login
